chore(virtctl): Move remove volume command to own file
